feat(minifier): substitute switch with one stmt whose parent is not block
